Added login.php page
login.php was created to handle the form submit.
It was taken from ch12, script 12.3.

I sha1 encrypted the user_id cookie, for security purposes.
I was unable to get the sha1 function to work directly
with the  array value, so I made a variable
for it to work on.  This seems to have solved the problem.