Created logout.php and made nav login dynamic.
Logout.php clears the two cookies that were created by changing
their values to empty strings.

Updated the nav in /includes/header.php so that it will
display login or logout dependent on the presence of the
user_id cookie.  The if statement checks to make sure that
the page you are on is not the logout page.  This is because
the cookie exists on the logout page, but you need it to say
login there.