created the forum.php page
Based mainly on the 17.4 script, it has been modified to use forum_id
instead of lang_id (which I killed off long ago).  The query used is
bananas.  I hope someday to be able to write a query that complex.
While I understand what each of the parts do, I am unsure I would
be ballsy enough to just take a blind stab at something like that.

I decided to keep the tables for displaying the messages, but I am not
possitive that would be a real solution if I had more time to mess with
this.