updated header.php logout bug
logout link was showing in the header on the logout.php page.

This was easily fixed by grabbing the old logout cookie code
and implementing it in the add log out link portion of the
header.  It's a if with isset for the cookie, but also checks to
make sure you are not on the logout page.