created post_form.php
This form is taken from 17.6.  Modified the initial search for ,
as I am not using that functionality.  Instead, I created a variable
named  in the header.  I look for that now, and if it doesn't
exist, then I know that the page is being accessed directly. (I had
thought about creating a session variable, but realized you could
access directly after you created the session, so that would defeat
the purpose.  Otherwise, pretty much straight from the book.