added the post.php page.
Finally made it all the way through the chapter.  Now I
can see the form for submitting a new thread, but have unearthed
other issues.

I can't post to any of the forums.  I am not sure what the issue is
yet, but
i know it has to do with the header and the fid not staying static.
Maybe I have to make a session variable for it?  Not sure.

That's tomorrows issue!